split deps.ts into dependencies submodule
